+++ libraries/M5ROTATE8/m5rotate8.cpp
@@ -1,22 +1,28 @@
 //
 //    FILE: m5rotate8.cpp
 //  AUTHOR: Rob Tillaart
-// VERSION: 0.3.0
+// VERSION: 0.4.0
 // PURPOSE: Arduino library for M5 8ROTATE 8x rotary encoders
 //     URL: https://github.com/RobTillaart/M5ROTATE8
 
 
 #include "m5rotate8.h"
 
+//  FIRMWARE V1 REGISTERS
+#define M5ROTATE8_REG_ADDRESS               0xFF
+#define M5ROTATE8_REG_VERSION               0xFE
+#define M5ROTATE8_REG_BASE_ABS              0x00
+#define M5ROTATE8_REG_BASE_REL              0x20
+#define M5ROTATE8_REG_BASE_RESET            0x40
+#define M5ROTATE8_REG_BASE_BUTTON_VALUE     0x50
+#define M5ROTATE8_REG_SWITCH                0x60
+#define M5ROTATE8_REG_RGB                   0x70
 
-#define M5ROTATE8_REG_ADDRESS            0xFF
-#define M5ROTATE8_REG_VERSION            0xFE
-#define M5ROTATE8_REG_BASE_ABS           0x00
-#define M5ROTATE8_REG_BASE_REL           0x20
-#define M5ROTATE8_REG_BASE_RESET         0x40
-#define M5ROTATE8_REG_BASE_BUTTON        0x50
-#define M5ROTATE8_REG_SWITCH             0x60
-#define M5ROTATE8_REG_RGB                0x70
+//  FIRMWARE V2 REGISTERS
+#define M5ROTATE8_REG_BASE_BUTTON_TOGGLE    0x58
+#define M5ROTATE8_REG_ENCODER_MASK          0x61
+#define M5ROTATE8_REG_BUTTON_MASK           0x62
+
 
 
 M5ROTATE8::M5ROTATE8(uint8_t address, TwoWire *wire)
@@ -42,6 +48,7 @@ bool M5ROTATE8::isConnected()
 
 bool M5ROTATE8::setAddress(uint8_t address)
 {
+  if ((address < 8) || (address > 119)) return false;
   _address = address;
   write8(M5ROTATE8_REG_ADDRESS, _address);
   return isConnected();
@@ -60,7 +67,6 @@ uint8_t M5ROTATE8::getVersion()
 }
 
 
-////////////////////////////////////////////////
 //
 //  ROTARY ENCODER PART
 //
@@ -88,7 +94,7 @@ bool M5ROTATE8::getKeyPressed(uint8_t channel)
   {
     return false;
   }
-  return (0 == read8(M5ROTATE8_REG_BASE_BUTTON + channel));
+  return (0 == read8(M5ROTATE8_REG_BASE_BUTTON_VALUE + channel));
 }
 
 
@@ -112,7 +118,6 @@ void M5ROTATE8::resetAll()
 }
 
 
-////////////////////////////////////////////////
 //
 //  INPUT SWITCH PART
 //
@@ -122,7 +127,6 @@ uint8_t M5ROTATE8::inputSwitch()
 }
 
 
-////////////////////////////////////////////////
 //
 //  LED PART
 //
@@ -134,6 +138,12 @@ bool M5ROTATE8::writeRGB(uint8_t channel, uint8_t R, uint8_t G, uint8_t B)
   }
   write24(M5ROTATE8_REG_RGB + (channel * 3), R, G, B);
   return true;
+}
+
+
+uint32_t M5ROTATE8::readRGB(uint8_t channel)
+{
+  return read24(M5ROTATE8_REG_RGB + (channel * 3));
 }
 
 
@@ -153,7 +163,47 @@ bool M5ROTATE8::allOff()
 }
 
 
-////////////////////////////////////////////////
+
+//
+//  FIRMWARE V2
+//
+bool M5ROTATE8::setButtonToggleCount(uint8_t channel, uint8_t value)
+{
+  if (channel > 7)
+  {
+    return false;
+  }
+  return write8(M5ROTATE8_REG_BASE_BUTTON_TOGGLE + channel, value);
+}
+
+
+uint8_t M5ROTATE8::getButtonToggleCount(uint8_t channel)
+{
+  if (channel > 7)
+  {
+    return 0;
+  }
+  return read8(M5ROTATE8_REG_BASE_BUTTON_TOGGLE + channel);
+}
+
+
+//  0 = no change, 1 = changed
+uint8_t M5ROTATE8::getEncoderChangeMask()
+{
+  return read8(M5ROTATE8_REG_ENCODER_MASK);
+}
+
+
+//  0 = not pressed, 1 = pressed (inverted the datasheetV2 specification)
+//  seems more logical
+uint8_t M5ROTATE8::getButtonChangeMask()
+{
+  //  invert register to be more logical IMHO.
+  return read8(M5ROTATE8_REG_BUTTON_MASK) ^ 0xFF;
+}
+
+
+//////////////////////////////////////////////////////////////////////////////
 //
 //  PRIVATE
 //
@@ -195,6 +245,31 @@ bool M5ROTATE8::write24(uint8_t reg, uint8_t R, uint8_t G, uint8_t B)
   _wire->write(B);
   _error = _wire->endTransmission();
   return (_error == 0);
+}
+
+
+uint32_t M5ROTATE8::read24(uint8_t reg)
+{
+  _wire->beginTransmission(_address);
+  _wire->write(reg);
+  _error = _wire->endTransmission();
+  if (_error != 0)
+  {
+    //  error handling
+    return 0;
+  }
+  if (_wire->requestFrom(_address, (uint8_t)3) != 3)
+  {
+    //  error handling
+    return 0;
+  }
+  uint32_t value = 0;
+  value += _wire->read();
+  value <<= 8;
+  value += _wire->read();
+  value <<= 8;
+  value += _wire->read();
+  return value;
 }