+++ libraries/M5ROTATE8/CHANGELOG.md
@@ -6,6 +6,20 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
 
+## [0.4.0] 2024-06-15
+- add **uint32_t readRGB(uint8_t channel)**
+- add firmware V2 functions
+  - **void setButtonToggleCount(uint8_t channel, uint8_t value = 0)**
+  - **uint8_t getButtonToggleCount(uint8_t channel)**
+  - **uint8_t getButtonRegister()**
+  - **uint8_t getEncoderChangeFlag()**
+- improved **setAddress()**
+- add examples
+- update readme.md
+- minor edits
+
+----
+
 ## [0.3.0] 2023-12-05
 - refactor API
 - update readme